First part of the wall
So far this is just 2/3 basic tables.
side ones can list other friends and for the other we can do whatever
(can be removed).
Center table is dynamicly sized to the page and should show the posts to
the user's wall. The person posting. it's karma and i put a table spot
for the like/dislike button (will end up being a drop down list)
I need to add a thing to make a post at the top.

So far this does not differentiate who's looking at which wall